refactor sobre la view
